Extract shared function "pad" from positional_printer.hpp to some utils
